+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (fixes [#37](https://github.com/flipp-oss/deimos/issues/37))
 - Raise error if producing without a topic
   (fixes [#50](https://github.com/flipp-oss/deimos/issues/50))
-
+- Don't try to load producers/consumers when running rake tasks involving webpacker or assets
 
 ## 1.8.2-beta2 - 2020-09-15